Add style for the index.html
For all the odd numbers, the background color is light pink, and the
font is Verdana.

For the second element, the text color set to blue with font size 36px.

For the third element, besides style added to the odd number, the font
size set to 24px.

For the fourth element, the background color is light green, and font is
24px and bold.